+++ src/table/TableHeadWrapper.jsx
@@ -29,6 +29,7 @@ export default function TableHeadWrapper({
   changeSortOrder,
   constraints,
   setFocusedCell,
+  page,
 }) {
   const classes = useStyles(getHeadStyle(layout, theme));
 
@@ -52,7 +53,8 @@ export default function TableHeadWrapper({
                   setFocusedCell,
                   changeSortOrder,
                   layout,
-                  column.isDim
+                  column.isDim,
+                  page
                 )
               }
             >
@@ -80,4 +82,5 @@ TableHeadWrapper.propTypes = {
   changeSortOrder: PropTypes.func.isRequired,
   constraints: PropTypes.object.isRequired,
   setFocusedCell: PropTypes.func.isRequired,
+  page: PropTypes.number.isRequired,
 };
