+++ src/table/TableWrapper.jsx
@@ -66,7 +66,7 @@ export default function TableWrapper(props) {
     >
       <TableContainer className={classes[containerMode]}>
         <Table stickyHeader aria-label="sticky table">
-          <TableHeadWrapper {...props} focusedCell={focusedCell} setFocusedCell={setFocusedCell} />
+          <TableHeadWrapper {...props} page={page} focusedCell={focusedCell} setFocusedCell={setFocusedCell} />
           <TableBodyWrapper {...props} page={page} focusedCell={focusedCell} setFocusedCell={setFocusedCell} />
         </Table>
       </TableContainer>
