+++ src/table/cells/withSelections.jsx
@@ -1,28 +1,17 @@
 import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
-import handleCellFocus from './handle-cell-focus';
+import { handleCellFocus } from './handle-cell-focus';
 import { selectCell } from '../selections-utils';
 import { getSelectionStyle } from '../styling-utils';
 
 export default function withSelections(CellComponent) {
   const HOC = (props) => {
-    const {
-      columnIndex,
-      cell,
-      selState,
-      selDispatch,
-      styling,
-      focusedCell,
-      setFocusedCell,
-      rootElement,
-      pageInfo,
-      page,
-    } = props;
+    const { cell, selState, selDispatch, styling, focusedCell, setFocusedCell, rootElement, page } = props;
     const handleMouseUp = (evt) => {
       cell.isDim &&
         evt.button === 0 &&
-        handleCellFocus(columnIndex, cell, focusedCell, setFocusedCell, rootElement, pageInfo, page) &&
-        selectCell(cell, selState, selDispatch, focusedCell, setFocusedCell, rootElement, pageInfo, page, evt);
+        handleCellFocus(cell, focusedCell, setFocusedCell, rootElement, page) &&
+        selectCell(cell, selState, selDispatch, evt);
     };
 
     const selectionStyling = useMemo(() => getSelectionStyle(styling, cell, selState), [styling, cell, selState]);
@@ -31,7 +20,6 @@ export default function withSelections(CellComponent) {
   };
 
   HOC.propTypes = {
-    columnIndex: PropTypes.number.isRequired,
     styling: PropTypes.object.isRequired,
     cell: PropTypes.object.isRequired,
     selState: PropTypes.object.isRequired,
@@ -39,7 +27,6 @@ export default function withSelections(CellComponent) {
     focusedCell: PropTypes.array.isRequired,
     setFocusedCell: PropTypes.func.isRequired,
     rootElement: PropTypes.object.isRequired,
-    pageInfo: PropTypes.object.isRequired,
     page: PropTypes.number.isRequired,
   };
